Swtorzylem CartController: getAllProductsFromCart - zwraca liste produktow z danego koszyka createCart - tworzy pusty koszyk addProductToCart - dodaje produkt do koszyka createOrderFromCart - tworzy order z danego koszyka deleteProductFromCart - usuwa produkt z koszyka
W OrderResponse zmienilem typ totalPrice z Double na BigDecimal

+++ src/main/java/com/kodilla/ecommercee/dto/response/OrderResponse.java
@@ -1,4 +1,6 @@
 package com.kodilla.ecommercee.dto.response;
 
-public record OrderResponse (Long id,Long cardId, Long userId, Double totalPrice, String status){
+import java.math.BigDecimal;
+
+public record OrderResponse(Long id, Long cardId, Long userId, BigDecimal totalPrice, String status) {
 }
